+++ R/mcnemar.R
@@ -1,15 +1,22 @@
+#' @name mcnemar_test_pv
+#'
 #' @title
 #' McNemar's Test for Count Data
 #'
 #' @description
 #' Performs McNemar's chi-square test or an exact variant to assess the symmetry
 #' of rows and columns in a 2-by-2 contingency table. In contrast to
-#' [stats::mcnemar.test()], it is vectorised, only calculates p-values and
+#' [`stats::mcnemar.test()`], it is vectorised, only calculates p-values and
 #' offers their exact computation. Furthermore, it is capable of returning the
 #' discrete p-value supports, i.e. all observable p-values under a null
 #' hypothesis. Multiple tables can be analysed simultaneously. In two-sided
 #' tests, several procedures of obtaining the respective p-values are
-#' implemented. It is a special case of the [binomial test][binom.test.pv()].
+#' implemented. It is a special case of the [binomial test][binom_test_pv()].
+#'
+#' `r lifecycle::badge('deprecated')`\cr
+#' **Note**: Please use `mcnemar_test_pv()`! The older `mcnemar.test.pv()` is
+#' deprecated in order to migrate to snake case. It will be removed in a future
+#' version.
 #'
 #' @param x   integer vector with four elements, a 2-by-2 matrix or an integer
 #'            matrix (or data frame) with four columns where each line
@@ -17,24 +24,23 @@
 #'
 #' @template param
 #' @templateVar alternative TRUE
-#' @templateVar ts.method FALSE
 #' @templateVar exact TRUE
 #' @templateVar correct TRUE
-#' @templateVar simple.output TRUE
+#' @templateVar simple_output TRUE
 #'
 #' @details
 #' The parameters `x` and `alternative` are vectorised. They are replicated
 #' automatically, such that the number of `x`'s rows is the same as the length
-#' of `alternative`. This allows multiple hypotheses to be tested
+#' of `alternative`. This allows multiple null hypotheses to be tested
 #' simultaneously. Since `x is (if necessary) coerced to a matrix with four
 #' columns, it is replicated row-wise.
 #'
 #' It can be shown that McNemar's test is a special case of the binomial test.
-#' Therefore, its computations are handled by [binom.test.pv()]. In contrast to
-#' that function, `mcnemar.test.pv` does not allow specifying exact two-sided
-#' p-value calculation procedures. The reason is that McNemar's exact test
-#' always tests for a probability of 0.5, in which case all these exact
-#' two-sided p-value computation methods yield exactly the same results.
+#' Therefore, its computations are handled by [`binom_test_pv()`]. In
+#' contrast to that function, `mcnemar_test_pv()` does not allow specifying
+#' exact two-sided p-value calculation procedures. The reason is that McNemar's
+#' exact test always tests for a probability of 0.5, in which case all these
+#' exact two-sided p-value computation methods yield exactly the same results.
 #'
 #' @template return
 #'
@@ -43,7 +49,7 @@
 #'   Wiley & Sons. pp. 411–413. \doi{10.1002/0471249688}
 #'
 #' @seealso
-#' [stats::mcnemar.test()], [binom.test.pv()]
+#' [`stats::mcnemar.test()`], [`binom_test_pv()`]
 #'
 #' @examples
 #' # Constructing
@@ -56,29 +62,29 @@
 #' df <- data.frame(S1, F1, S2, F2)
 #'
 #' # Computation of exact p-values and their supports
-#' results.ex  <- mcnemar.test.pv(df)
-#' raw.pvalues <- results.ex$get_pvalues()
-#' pCDFlist    <- results.ex$get_pvalue_supports()
+#' results_ex  <- mcnemar_test_pv(df)
+#' raw_pvalues <- results_ex$get_pvalues()
+#' pCDFlist    <- results_ex$get_pvalue_supports()
 #'
 #' # Computation of chisquare p-values and their supports
-#' results.cs  <- mcnemar.test.pv(df, exact = FALSE)
-#' raw.pvalues <- results.cs$get_pvalues()
-#' pCDFlist    <- results.cs$get_pvalue_supports()
+#' results_cs  <- mcnemar_test_pv(df, exact = FALSE)
+#' raw_pvalues <- results_cs$get_pvalues()
+#' pCDFlist    <- results_cs$get_pvalue_supports()
 #'
 #' @importFrom stats pchisq
 #' @importFrom checkmate assert_integerish
 #' @export
-mcnemar.test.pv <- function(
+mcnemar_test_pv <- function(
   x,
   alternative = "two.sided",
   exact = TRUE,
   correct = TRUE,
-  simple.output = FALSE
+  simple_output = FALSE
 ) {
   # plausibility checks of input parameters
 
   # define error message for malformed x
-  error.msg.x <- paste("'x' must either be a 2-by-2 matrix,",
+  error_msg_x <- paste("'x' must either be a 2-by-2 matrix,",
                        "a four-element vector or a four-column matrix")
 
   #  if x is a vector, make it a matrix with one row
@@ -88,7 +94,7 @@ mcnemar.test.pv <- function(
   if(is.data.frame(x))
     x <- as.matrix(x)
   # if x is a list, then abort
-  if(is.list(x)) stop(error.msg.x)
+  if(is.list(x)) stop(error_msg_x)
   # when x is a matrix, it must satisfy some conditions
   if(is.matrix(x)) {
     # check if all values are non-negative and close to integer
@@ -97,58 +103,55 @@ mcnemar.test.pv <- function(
     x <- round(x)
     # stop immediately, if dimensions are violated
     if(any(dim(x) != c(2, 2)) && ncol(x) != 4 && nrow(x) != 4)
-      stop(error.msg.x)
+      stop(error_msg_x)
     # 2-by-2 matrices are transformed to single-row matrix
     if(all(dim(x) == c(2, 2))) {
       x <- matrix(as.vector(x), 1, 4,
         dimnames = list(NULL,
-          make.names(paste(rep(colnames(x), rep(2,2)), rownames(x)))
+          make.names(paste(rep(colnames(x), rep(2, 2)), rownames(x)))
         )
       )
     } else
       # transpose 4-row matrix (with more or less columns than 4) to 4-column matrix
       if((nrow(x) == 4 && ncol(x) != 4))
         x <- t(x)
-  } else stop(error.msg.x)
+  } else stop(error_msg_x)
 
   # lengths
-  len.x <- nrow(x)
-  len.a <- length(alternative)
-  len.g <- max(len.x, len.a)
+  len_x <- nrow(x)
+  len_a <- length(alternative)
+  len_g <- max(len_x, len_a)
 
   # check exactness and continuity correction arguments
   qassert(exact, "B1")
   if(!exact) qassert(correct, "B1")
 
   # check alternatives
-  for(i in seq_len(len.a)){
+  for(i in seq_len(len_a)){
     alternative[i] <- match.arg(
       alternative[i],
       c("two.sided", "less", "greater")
     )
   }
-  if(len.a < len.g) alternative <- rep_len(alternative, len.g)
+  if(len_a < len_g) alternative <- rep_len(alternative, len_g)
 
   # enlarge input matrix if necessary
-  if(len.x < len.g)
-    x <- matrix(rep_len(as.vector(t(x)), 4 * len.g), len.g, 4, TRUE)
+  if(len_x < len_g)
+    x <- matrix(rep_len(as.vector(t(x)), 4 * len_g), len_g, 4, TRUE)
 
   # check output type argument
-  qassert(simple.output, "B1")
+  qassert(simple_output, "B1")
 
   # test parameters
   b      <- x[, 2]
   c      <- x[, 3]
   n      <- b + c
-  params <- unique(data.frame(n, alternative))
-  len.u  <- nrow(params)
-  alt.u  <- params$alternative
 
   # compute test results
-  res <- binom.test.pv(b, n, 0.5, alternative, "central", exact, correct, simple.output)
+  res <- binom_test_pv(b, n, 0.5, alternative, "central", exact, correct, simple_output)
 
   # create output object
-  out <- if(!simple.output) {
+  out <- if(!simple_output) {
     if(is.null(colnames(x)))
       colnames(x) <- paste0("x", c("[1, 1]", "[2, 1]", "[1, 2]", "[2, 2]"))
 
@@ -167,7 +170,7 @@ mcnemar.test.pv <- function(
         list(
           observations = as.data.frame(x),
           nullvalues = data.frame(
-            `counter-diagonal values proportions` = rep(0.5, len.g),
+            `counter-diagonal values proportion` = rep(0.5, len_g),
             check.names = FALSE
           ),
           parameters = data.frame(
@@ -186,4 +189,19 @@ mcnemar.test.pv <- function(
 
   # return results
   return(out)
+}
+
+#' @rdname mcnemar_test_pv
+#' @export
+#' @importFrom lifecycle deprecate_soft
+mcnemar.test.pv <- function(
+    x,
+    alternative = "two.sided",
+    exact = TRUE,
+    correct = TRUE,
+    simple.output = FALSE
+) {
+  deprecate_soft("0.2", "mcnemar.test.pv()", "mcnemar_test_pv()")
+
+  mcnemar_test_pv(x, alternative, exact, correct, simple.output)
 }